+++ CHANGELOG.md
@@ -11,6 +11,9 @@
   - Zettelkasten directory is now provided via the `-d` argument.
     - Its default, `~/zettelkasten`, is used when not specified.
     - This directory must exist, otherwise neuron will error out.
+  - The output directory is now moved to `.neuron/output` under the Zettelkasten directory.
+  - Added `neuron open` to open the locally generated Zettelkasten site.
+  - `neuron ... rib serve` is now `neuron rib -wS`.
 
 ## 0.2.0.0